requirements.txt
attrs==24.2.0
colorama==0.4.6
Flask==3.0.3
flatbuffers==24.3.25
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.28.3
pybase64==1.4.0
PyMatting==1.1.12
pyreadline3==3.5.4
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rpds-py==0.20.0
scikit-image==0.24.0
scipy==1.14.1
sympy==1.13.3
tifffile==2024.9.20
tqdm==4.66.5
urllib3==2.2.3
Werkzeug==3.0.4
